renaming loop_reorder_at_models
