+++ tests/loop/reorder_at/loop_reorder_at_novalues.ml
@@ -1,17 +1,14 @@
-
+(* Deprecated *)
 open Optitrust
 open Target
 
 (* let _ = Flags.check_validity := true *)
-let _ = Flags.typechecking_mode := Flags.ProofPreserving
-let _ = Flags.recompute_resources_between_steps := true
+(* let _ = Flags.recompute_resources_between_steps := true *)
 
 let _ = Run.script_cpp (fun _ ->
   (* TODO: minimize contracts *)
   !! Loop.reorder_at ~order:["c";"b"] [cFunBody "f1"; cForBody "c"; dSeqNth 0];
   !! Loop.reorder_at ~order:["k";"i";"j"] [cFunBody "f2"; cForBody "k"; sInstr "sum +="];
-  (* TODO:
   !! Trace.failure_expected (fun _e -> true) (fun () ->
     Loop.reorder_at ~order:["c";"b"] [cFunBody "f1_wrong"; sInstr "x += c"]);
-    *)
 )
