+++ tests/resources_with_models/computation/optitrust_header.ml
@@ -1,9 +1,3 @@
-(* Deprecated *)
 open Optitrust
-
-(* let _ = Flags.check_validity := true *)
-(* let _ = Flags.recompute_resources_between_steps := true *)
-(*let _ = Flags.debug_parsing_serialization := true*)
-(*let _ = Flags.resource_errors_as_warnings := true*)
 
 let _ = Run.script_cpp ~inline:["../../../include/optitrust_models.h"] (fun () -> ())
